% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_time_series.R
\name{assign_indices}
\alias{assign_indices}
\title{Assign Basin and Bridge Indices}
\usage{
assign_indices(states)
}
\arguments{
\item{states}{State classifications}
}
\value{
List with basin_index and bridge_index vectors
}
\description{
Assign Basin and Bridge Indices
}
