% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segmentation.R
\name{create_ggplot_visualizations}
\alias{create_ggplot_visualizations}
\title{Create ggplot2 Visualizations}
\usage{
create_ggplot_visualizations(segmentation_result, plot_type)
}
\arguments{
\item{segmentation_result}{Segmentation results}

\item{plot_type}{Type of plot to create}
}
\value{
ggplot object or list of ggplot objects
}
\description{
Create ggplot2 Visualizations
}
