% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{find_next_basin}
\alias{find_next_basin}
\title{Find Next Basin}
\usage{
find_next_basin(data, current_index)
}
\arguments{
\item{data}{Segmentation data}

\item{current_index}{Current row index}
}
\value{
Basin index of next basin, or NA if none found
}
\description{
Find Next Basin
}
