% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{find_previous_basin}
\alias{find_previous_basin}
\title{Find Previous Basin}
\usage{
find_previous_basin(data, current_index)
}
\arguments{
\item{data}{Segmentation data}

\item{current_index}{Current row index}
}
\value{
Basin index of previous basin, or NA if none found
}
\description{
Find Previous Basin
}
