% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{summarize_classification}
\alias{summarize_classification}
\title{Summarize Classification Results}
\usage{
summarize_classification(classified_data)
}
\arguments{
\item{classified_data}{Data frame with state classifications}
}
\value{
Summary of classification results
}
\description{
Summarize Classification Results
}
