% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_time_series.R
\name{summarize_segmentation}
\alias{summarize_segmentation}
\title{Summarize Segmentation Results}
\usage{
summarize_segmentation(result_data)
}
\arguments{
\item{result_data}{Segmentation result data frame}
}
\value{
Summary statistics
}
\description{
Summarize Segmentation Results
}
